+++ packages/tupaia-web/src/features/Visuals/Matrix/index.ts
@@ -1,0 +1,5 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+export { Matrix } from './Matrix';
